+++ MyWatch/src/sensors/AudioLevelListener.java
@@ -4,5 +4,5 @@ public interface AudioLevelListener {
 
 	public void onVeryLoudSoundDetected();	
 	
-	public void onLevelChanged(double value);
+	public void onLevelChanged(int value);
 }
